backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
bcrypt
celery
redis
minio
google-genai
pypdf
prometheus-fastapi-instrumentator
python-multipart
alembic
pytest
httpx
pytest-asyncio
pytest-cov
email-validator
